Use a cluster-compatible MOVED response.
This is just a quick hack to get going.

We'll probably need to fetch the keys and compute a real hash slot even
if we don't do partitioning, to avoid confusing clients that care about
that.

We should also consider the impact of this in MULTI/EXEC and potentially
other specific scenarios.