Add RESET as unsupported.
RESET cannot intercepted and applied through RedisRaft, because it needs
to operate on the client connection itself.

On the other hand, if it is not intercepted it will not clear the MULTI
state maintained inside RedisRaft.

In the future we may want to offload MULTI state handling back to Redis,
but this will require some module API support to be able to bundle the
commands in a single entry.